Move AnnotationDock
